test(workflow-core): consolidate PortIdentity key serde specs
